Sync

+++ PostModule/src/main/java/com/meminator/postmodule/Services/PostService.java
@@ -4,10 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.meminator.postmodule.DAO.PostDAO;
 import com.meminator.postmodule.DAO.RegisteredUserDAO;
 import com.meminator.postmodule.DAO.TagDAO;
-import com.meminator.postmodule.Models.Post;
-import com.meminator.postmodule.Models.PostVM;
-import com.meminator.postmodule.Models.RegisteredUser;
-import com.meminator.postmodule.Models.Tag;
+import com.meminator.postmodule.Models.*;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -56,19 +53,63 @@ public class PostService {
         this.restTemplate = restTemplate;
     }
 
-    public Post createPost(Post post, String username){
+    public Post createPost(PostVM postVM, String username){
 
-        if(post.getImageURL().equals("")){
+        Optional<RegisteredUser> user = registeredUserDAO.getUser(username);
+        
+        if(postVM.getRepostID() != 0){
+            Optional<Post> repost = postDAO.getPost(postVM.getRepostID());
+            if(repost.isPresent()){
+
+                Post tmp = repost.get();
+                Post post = new Post();
+                post.setUser(user.get());
+                post.setInfo(postVM.getInfo());
+                post.setImageID(tmp.getImageID());
+                post.setImageURL(tmp.getImageURL());
+                post.setRepost(tmp);
+                List<Tag> tags = tagDAO.getAllInList(postVM.getTags());
+                for(Tag i: tags){
+                    if(postVM.getTags().contains(i.getName())) postVM.getTags().remove(i.getName());
+                }
+                for(String i: postVM.getTags()){
+                    tags.add(tagDAO.createTag(new Tag(i)));
+                }
+                post.setTags(tags);
+                Post newPost =  postDAO.savePost(post);
+                if(newPost != null){
+                    asyncSender.sendPost(new PostVMS(post));
+                }
+                return newPost;
+
+            }else{
+                throw new IllegalArgumentException("Post doesn't exist");
+            }
+        }
+
+
+        if(postVM.getImageURL().equals("")){
             throw new IllegalArgumentException("Url of an image isn't specified!");
         }
 
-        Optional<RegisteredUser> user = registeredUserDAO.getUser(username);
         if(user.isPresent()){
+            Post post = new Post();
             post.setUser(user.get());
-
+            post.setInfo(postVM.getInfo());
+            post.setImageURL(postVM.getImageURL());
+            Optional<Post> repost = postDAO.getPost(postVM.getRepostID());
+            post.setRepost(null);
+            List<Tag> tags = tagDAO.getAllInList(postVM.getTags());
+            for(Tag i: tags){
+                if(postVM.getTags().contains(i.getName())) postVM.getTags().remove(i.getName());
+            }
+            for(String i: postVM.getTags()){
+                tags.add(tagDAO.createTag(new Tag(i)));
+            }
+            post.setTags(tags);
             Post newPost =  postDAO.savePost(post);
             if(newPost != null){
-                asyncSender.sendPost(new PostVM(post));
+                asyncSender.sendPost(new PostVMS(post));
             }
             return newPost;
         }else{
@@ -128,8 +169,15 @@ public class PostService {
         }
     }
 
-    public boolean deletePost(Long id){
-        return postDAO.deletePost(id);
+    public boolean deletePost(Long id, String username){
+        Optional<Post> post = postDAO.getPost(id);
+        if(post.isPresent())
+            if(post.get().getUser().getUsername().equals(username)) {
+                asyncSender.deletePost(id);
+                return postDAO.deletePost(id);
+            }
+            else throw new SecurityException("User cannot delete this post!");
+        else throw new IllegalArgumentException("Post with given id does not exist!");
     }
 
     public Post editPost(Long id, Post newpost){
@@ -155,11 +203,10 @@ public class PostService {
         JSONObject request = new JSONObject();
         request.put("username", username);
         List<String> response = restTemplate.postForObject(
-                "http://userModule/follow/myFriends",
+                "http://userModule/myFriends",
                 username,
                 List.class
                 );
-            System.out.println(response.get(1));
             List<RegisteredUser> users = registeredUserDAO.findAllByUsernames(response);
             return postDAO.getPostByFollowers(users);
     }
